moving reports logic to a report service
